requirements.txt
fastapi
uvicorn
pydantic
google-generativeai>=0.5.4
